Disable affinitiy
The current approach was intended to schedule a VM alongside the vmctl pod.
However, the issue is that it does not work.

Currently the individual VMs in different controllers (i.e. Deployment) don't
have individual labels, which prevents us frm using affinity rules to pin
a VM to it's vmctl pod.

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>